pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "codetandem"
version = "0.1.0"
description = "A CLI tool for AI-powered collaborative coding"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer[all]>=0.9.0",
    "keyring>=24.0.0",
    "rich>=13.0.0",
    "markdown>=3.4.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "anthropic>=0.8.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
]

[project.scripts]
codetandem = "codetandem.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
